LSF bug fixed: large strip size

+++ Typedefine.h
@@ -102,8 +102,8 @@ typedef struct tagTransParam
 
 typedef struct tagCSize
 {
-	unsigned int width;
-	unsigned int height;
+	long int width;
+	long int height;
 } CSize;
 
 typedef struct tagNCCflag
